+++ csvkit/convert/fixed.py
@@ -2,9 +2,9 @@
 
 from codecs import iterdecode
 from collections import namedtuple
+from io import StringIO
 
 import agate
-import six
 
 
 def fixed2csv(f, schema, output=None, skip_lines=0, **kwargs):
@@ -27,10 +27,10 @@ def fixed2csv(f, schema, output=None, skip_lines=0, **kwargs):
     :param skip_lines:
         The number of lines to skip from the top of the file.
     """
-    streaming = True if output else False
+    streaming = bool(output)
 
     if not streaming:
-        output = six.StringIO()
+        output = StringIO()
 
     try:
         encoding = kwargs['encoding']
@@ -59,7 +59,7 @@ def fixed2csv(f, schema, output=None, skip_lines=0, **kwargs):
     return ''
 
 
-class FixedWidthReader(six.Iterator):
+class FixedWidthReader:
     """
     Given a fixed-width file and a schema file, produce an analog to a csv
     reader that yields a row of strings for each line in the fixed-width file,
@@ -95,7 +95,7 @@ class FixedWidthReader(six.Iterator):
 FixedWidthField = namedtuple('FixedWidthField', ['name', 'start', 'length'])
 
 
-class FixedWidthRowParser(object):
+class FixedWidthRowParser:
     """
     Instantiated with a schema, able to return a sequence of trimmed strings
     representing fields given a fixed-length line. Flexible about where the
@@ -135,7 +135,7 @@ class FixedWidthRowParser(object):
         return [field.name for field in self.fields]
 
 
-class SchemaDecoder(object):
+class SchemaDecoder:
     """
     Extracts column, start, and length columns from schema rows. Once
     instantiated, each time the instance is called with a row, a