+++ csvkit/utilities/csvjson.py
@@ -1,19 +1,12 @@
 #!/usr/bin/env python
 
-import codecs
 import datetime
 import decimal
+import json
 import sys
-
-try:
-    import json
-    from collections import OrderedDict
-except ImportError:
-    from ordereddict import OrderedDict
-    import simplejson as json
+from collections import OrderedDict
 
 import agate
-import six
 
 from csvkit.cli import CSVKitUtility, match_column_identifier
 
@@ -63,7 +56,7 @@ class CSVJSON(CSVKitUtility):
             help='Disable type inference (and --locale, --date-format, --datetime-format) when parsing CSV input.')
 
     def __init__(self, args=None, output_file=None):
-        super(CSVJSON, self).__init__(args, output_file)
+        super().__init__(args, output_file)
 
         self.validate_args()
 
@@ -73,11 +66,7 @@ class CSVJSON(CSVKitUtility):
         }
 
         # We need to do this stream dance here, because we aren't writing through agate.
-        if six.PY2:
-            self.stream = codecs.getwriter('utf-8')(self.output_file)
-            self.json_kwargs['encoding'] = 'utf-8'
-        else:
-            self.stream = self.output_file
+        self.stream = self.output_file
 
     def main(self):
         """
@@ -106,7 +95,7 @@ class CSVJSON(CSVKitUtility):
         def default(obj):
             if isinstance(obj, (datetime.date, datetime.datetime)):
                 return obj.isoformat()
-            elif isinstance(obj, decimal.Decimal):
+            if isinstance(obj, decimal.Decimal):
                 return str(obj)
             raise TypeError('%s is not JSON serializable' % repr(obj))