+++ tests/test_utilities/test_csvcut.py
@@ -1,12 +1,7 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
 
 import sys
-
-try:
-    from mock import patch
-except ImportError:
-    from unittest.mock import patch
+from unittest.mock import patch
 
 from csvkit.utilities.csvcut import CSVCut, launch_new_instance
 from tests.utils import ColumnsTests, CSVKitTestCase, EmptyFileTests, NamesTests
@@ -31,11 +26,17 @@ class TestCSVCut(CSVKitTestCase, ColumnsTests, EmptyFileTests, NamesTests):
             ['1', '3'],
         ])
 
+    def test_linenumbers(self):
+        self.assertRows(['-c', '1,3', '-l', 'examples/dummy.csv'], [
+            ['line_number', 'a', 'c'],
+            ['1', '1', '3'],
+        ])
+
     def test_unicode(self):
         self.assertRows(['-c', '1,3', 'examples/test_utf8.csv'], [
             ['foo', 'baz'],
             ['1', '3'],
-            ['4', u'ʤ'],
+            ['4', 'ʤ'],
         ])
 
     def test_with_gzip(self):