+++ tests/test_utilities/test_csvformat.py
@@ -1,14 +1,8 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
 
 import sys
-
-import six
-
-try:
-    from mock import patch
-except ImportError:
-    from unittest.mock import patch
+from io import StringIO
+from unittest.mock import patch
 
 from csvkit.utilities.csvformat import CSVFormat, launch_new_instance
 from tests.utils import CSVKitTestCase, EmptyFileTests, stdin_as_string
@@ -52,7 +46,7 @@ class TestCSVFormat(CSVKitTestCase, EmptyFileTests):
         ])
 
     def test_quotechar(self):
-        input_file = six.StringIO('a,b,c\n1*2,3,4\n')
+        input_file = StringIO('a,b,c\n1*2,3,4\n')
 
         with stdin_as_string(input_file):
             self.assertLines(['-Q', '*'], [
@@ -63,7 +57,7 @@ class TestCSVFormat(CSVKitTestCase, EmptyFileTests):
         input_file.close()
 
     def test_doublequote(self):
-        input_file = six.StringIO('a\n"a ""quoted"" string"')
+        input_file = StringIO('a\n"a ""quoted"" string"')
 
         with stdin_as_string(input_file):
             self.assertLines(['-P', '#', '-B'], [
@@ -74,7 +68,7 @@ class TestCSVFormat(CSVKitTestCase, EmptyFileTests):
         input_file.close()
 
     def test_escapechar(self):
-        input_file = six.StringIO('a,b,c\n1"2,3,4\n')
+        input_file = StringIO('a,b,c\n1"2,3,4\n')
 
         with stdin_as_string(input_file):
             self.assertLines(['-P', '#', '-U', '3'], [