+++ tests/test_utilities/test_csvsort.py
@@ -1,14 +1,8 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
 
 import sys
-
-import six
-
-try:
-    from mock import patch
-except ImportError:
-    from unittest.mock import patch
+from io import StringIO
+from unittest.mock import patch
 
 from csvkit.utilities.csvsort import CSVSort, launch_new_instance
 from tests.utils import ColumnsTests, CSVKitTestCase, EmptyFileTests, NamesTests, stdin_as_string
@@ -22,37 +16,31 @@ class TestCSVSort(CSVKitTestCase, ColumnsTests, EmptyFileTests, NamesTests):
             launch_new_instance()
 
     def test_runs(self):
-        if six.PY2:
-            self.get_output(['examples/test_utf8.csv'])
-        else:
-            self.assertRows(['examples/test_utf8.csv'], [
-                ['foo', 'bar', 'baz'],
-                ['1', '2', '3'],
-                ['4', '5', 'ʤ'],
-            ])
+        self.assertRows(['examples/test_utf8.csv'], [
+            ['foo', 'bar', 'baz'],
+            ['1', '2', '3'],
+            ['4', '5', 'ʤ'],
+        ])
 
     def test_encoding(self):
-        if six.PY2:
-            self.get_output(['-e', 'latin1', 'examples/test_latin1.csv'])
-        else:
-            self.assertRows(['-e', 'latin1', 'examples/test_latin1.csv'], [
-                ['a', 'b', 'c'],
-                ['1', '2', '3'],
-                ['4', '5', '©'],
-            ])
+        self.assertRows(['-e', 'latin1', 'examples/test_latin1.csv'], [
+            ['a', 'b', 'c'],
+            ['1', '2', '3'],
+            ['4', '5', '©'],
+        ])
 
     def test_sort_string_reverse(self):
         reader = self.get_output_as_reader(['-c', '1', '-r', 'examples/testxls_converted.csv'])
-        test_order = [u'text', u'Unicode! Σ', u'This row has blanks',
-                      u'Chicago Tribune', u'Chicago Sun-Times', u'Chicago Reader']
-        new_order = [six.text_type(r[0]) for r in reader]
+        test_order = ['text', 'Unicode! Σ', 'This row has blanks',
+                      'Chicago Tribune', 'Chicago Sun-Times', 'Chicago Reader']
+        new_order = [str(r[0]) for r in reader]
         self.assertEqual(test_order, new_order)
 
     def test_sort_date(self):
         reader = self.get_output_as_reader(['-c', '2', 'examples/testxls_converted.csv'])
-        test_order = [u'text', u'Chicago Tribune', u'Chicago Sun-Times',
-                      u'Chicago Reader', u'This row has blanks', u'Unicode! Σ']
-        new_order = [six.text_type(r[0]) for r in reader]
+        test_order = ['text', 'Chicago Tribune', 'Chicago Sun-Times',
+                      'Chicago Reader', 'This row has blanks', 'Unicode! Σ']
+        new_order = [str(r[0]) for r in reader]
         self.assertEqual(test_order, new_order)
 
     def test_no_blanks(self):
@@ -81,18 +69,18 @@ class TestCSVSort(CSVKitTestCase, ColumnsTests, EmptyFileTests, NamesTests):
 
     def test_no_inference(self):
         reader = self.get_output_as_reader(['--no-inference', '-c', '1', 'examples/test_literal_order.csv'])
-        test_order = [u'a', u'192', u'27', u'3']
-        new_order = [six.text_type(r[0]) for r in reader]
+        test_order = ['a', '192', '27', '3']
+        new_order = [str(r[0]) for r in reader]
         self.assertEqual(test_order, new_order)
 
     def test_sort_t_and_nulls(self):
         reader = self.get_output_as_reader(['-c', '2', 'examples/sort_ints_nulls.csv'])
         test_order = ['b', '1', '2', '']
-        new_order = [six.text_type(r[1]) for r in reader]
+        new_order = [str(r[1]) for r in reader]
         self.assertEqual(test_order, new_order)
 
     def test_stdin(self):
-        input_file = six.StringIO('a,b,c\n4,5,6\n1,2,3\n')
+        input_file = StringIO('a,b,c\n4,5,6\n1,2,3\n')
 
         with stdin_as_string(input_file):
             self.assertLines([], [